+++ models/demos/yolov8x/demo/images/bus.jpg
