+++ models/demos/yolov8x/tests/test_e2e_performant.py
@@ -9,7 +9,7 @@ import torch
 from loguru import logger
 
 from models.utility_functions import run_for_wormhole_b0
-from models.experimental.yolov8x.tests.yolov8x_e2e_performant import Yolov8xTrace2CQ
+from models.demos.yolov8x.tests.yolov8x_e2e_performant import Yolov8xTrace2CQ
 
 
 @run_for_wormhole_b0()
