+++ models/demos/yolov8x/tests/yolov8x_e2e_performant.py
@@ -3,7 +3,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import ttnn
-from models.experimental.yolov8x.tests.yolov8x_test_infra import create_test_infra
+from models.demos.yolov8x.tests.yolov8x_test_infra import create_test_infra
 
 try:
     pass
