+++ tests/scripts/single_card/run_single_card_demo_tests.sh
@@ -207,6 +207,13 @@ run_yolov8s_world_perf() {
 
 }
 
+run_yolov8x_perf() {
+
+  WH_ARCH_YAML=wormhole_b0_80_arch_eth_dispatch.yaml pytest -n auto --disable-warnings models/demos/yolov8x/demo/demo.py --timeout 600
+
+}
+
+
 main() {
   # For CI pipeline - source func commands but don't execute tests if not invoked directly
   if [[ "${BASH_SOURCE[0]}" != "${0}" ]]; then
